update for user space scheduler
