+++ sched/mysched.h
@@ -5,6 +5,7 @@
 // 1 for ascending, 0 for descending
 void swap(int * a, int * b);
 void sort(int value[], int index[], int N, int ascend);
+void resort(int value[], int index[], int N, int ascend, const int T[], const int T_inverse[]){
 
 void inverse_permutation(const int a[], int b[], int N);
 // flag == 1 for reading clock
